Grafieken wijzigingen Seppe

+++ MVC/Models/ChartDatasets.cs
@@ -19,6 +19,6 @@ namespace MVC.Models
     public string[] backgroundColor { get; set; }
     public string[] borderColor { get; set; }
     public string borderWidth { get; set; }
-    public int[] data { get; set; }
+    public double[] data { get; set; }
   }
 }
